requirements.txt
spotipy
openai
python-dotenv
Pillow
requests
pybase64
